requirements.txt
Django==3.1.4
djangorestframework==3.12.4
psycopg2==2.9.1
virtualenv==20.4.2
drf-yasg==1.20.0